+++ mblt_model_zoo/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.8.0"
+__version__ = "2.9.0"
 from . import utils, vision
 
 __all__ = ["utils", "vision"]